The F routine does not need an l parameter
The F^{x}(S) routine is only called with three different values for x:
the initialisation uses a fixed x=2, the main low-level routines use
x=l (the round number) and the finalisation uses x=2l. Rather than a
single routine that takes a parameter is better to create separate F_l
and F_2 routines. The loops in these routines have constant bounds and
so will be more likely to be easily unrolled by the compiler.