Support NORX08 and NORX16 variants
The only real changes were in the Initialisation routines (because the
key material for NORX16 is inserted into the state starting from S(2)
rather than S(4) for some reason) and in the Finalisation routines (as
the rate in NORX08 is smaller than the tag size, so additional work is
necessary).

This also required making some assumptions on relative sizes more
explicit in NORX32 and NORX64, via new generic parameters.