Add a "static" version of NORX6441 for proof
Using the SPARK tools on generic units is quite painful as SPARK will
only examine or prove generic units in the form of a specific
instantiation. My plan is to use this static version to work through the
subroutines one-by-one and when the problems have been solved the
solutions can be ported across to the generic package.