Split NORX.Compare_Tags into child unit
This allows different compiler flags to be applied. In particular we can
ensure that the optimiser is switched off for this function even if the
rest of the code is optimised. This prevents the code being compiled
into a logically identical form that short-circuits the comparison, and
so leaks some information on the similarity of the start of the tags via
timing differences.