Assert that full contents of arrays were written
The relevant procedures read or write data to array parameters using
index variables. These variables are initialised to the X'First when
they are declared, and the assertions that have been added prove that
at the end of the procedures they are equal to X'Last+1. That is, the
full array has been read or written.