Add a Key_Position param to the generic
The key material in NORX1641 is introduced starting at state position
S(2) rather than S(4), as it is in the other instantiations. Per P
Jovanovic this was a deliberate decision in order to keep the lower half
of the state initialised in a constant way, as the initial cryptanalysis
suggests the lower half may be slightly more exposed to differential
attacks.

Previously the different state position was silently triggered for all
16-bit instantiations of the NORX package, but it is probably better to
have an explicit parameter for this.