Extract tag from capacity, not rate
As per NORX 3.0 spec. Note that this means that for all defined variants
of NORX the tag can be extracted in one go, rather than potentially
having to extract it in multiple chunks with additional state
permutations in between. This complication only actually mattered for
NORX8, but getting rid of it is still a very welcome simplification, as
the Finalisation procedure was one of the most complex to get through
the SPARK verification process.