Eliminate Storage_Offset overflow risk in load_store preconditions
In SPARK Discovery 2017 (GPL) there is no longer an implicit assumption
that all Storage_Arrays are no longer than a Long_Long_Integer. This is
a problem as the index to Storage_Array is Storage_Offset, which is
defined by GNAT as being a signed integer - i.e. one bit wider than
Standard'Address_Size, which in practice means one bit wider than a
Long_Long_Integer.

There is therefore no machine integer that can represent the length of a
Storage_Array that runs from Storage_Offset'First to an index > 0. In
practice such an array can never arise, but it is now not possible to
refer to Storage_Array'Length in most contexts. Arithmetic on
Storage_Offset'First and Storage_Offset'Last is also liable to hit
overflow errors.

Instead it is necessary to do a check that the Storage_Array length will
fit into a Long_Long_Integer - a check that will always succeed - before
actually testing the length. Many subroutine preconditions have to be
re-written to do this, and ghost functions are used to make the results
more manageable. Thanks to Claire Dross from AdaCore for explanations.