Fix issue with indefinite public views of types and subroutine renaming
AARM 8.5.4.5/3 says: "If the renaming-as-body completes that declaration
before the subprogram it declares is frozen, the profile shall be mode
conformant with that of the renamed callable entity and the subprogram
it declares takes its convention from the renamed subprogram; otherwise,
the profile shall be subtype conformant with that of the renamed
callable entity and the convention of the renamed subprogram shall not
be Intrinsic."

Previously an indefinite type State(<>) was declared in the public part
of package NORX, and completed in the private part without a
discriminant, along with subroutines to operate on them.
NORX.Access_Internals then defined a subtype, also called State, in
order to define subroutines that had their bodies completed as renames
of those private subroutines.

This worked until GNAT Community 2018. It appears that GNAT Community
2018 has a stricter view of subtype conformance and does consider the
subtype Access_Internals.State of the public indefinite NORX.State(<>)
type to be subtype conformant with the the private definite NORX.State
that the private subroutines use. This might be because AARM 6.3.1.7/3
requires static matching of subtypes?

To be honest it is not very clear to me why this happens, because the
indefinite State type can only be instantiated as the matching definite
type, but the change should make little difference to anyone.

Signed-off-by: James Humphry <git@binecho.co.uk>